+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 #Change Log
 
+##v0.1.2
+* Fix help message
+* Add function for delete entry
+* Add command to delete tree
+* Add command to show server root dse
+
 ##v0.1.1
 * Search entry with filter
 * Show entry use full dn